govt-chatbot-app/backend/requirements.txt
flask==2.0.1
flask-cors==3.0.10
pymongo==4.0.1
python-dotenv==0.19.1
bson==0.5.10
requests==2.26.0 